WordCamp Base: Replace get_output_mode() with skip_stylesheet() to avoid bug
Previously (in r4590), Remote CSS's `get_output_mode()` was used, because it just pulls the theme mod setting directly, which is nice because it doesn't require Jetpack's Custom CSS module to be active on the site. That led to a bug where sometimes the CSS being output in the Site Cloner preview was wrong, though. If the target site has `replace` mode on, and the source site didn't, then the source site's theme stylesheet would be removed, because the Base theme didn't realize that it should be checking the mode on the target site, rather than the source site. Switching to `skip_stylesheet()` fixes that, since the `Source_Site_ID_Setting::preview_skip_stylesheet` hooks into the process and switches to the target site when the value is being pulled. Fixing that far outweighs the benefit of not requiring Custom CSS to be active.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/themes/wordcamp-base-v2@4781 74240141-8908-4e6f-9713-ba540dce6ec7